+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Throw error if invalid Payload on `CustomerGateway.create`
 * Allow payee ID to be passed in options params for transaction create
 * Add processor respone code and processor response text to authorization adjustments subfield in transaction response.
+* Add `merchant_id` alias to ConnectedMerchantStatusTransitioned and ConnectedMerchantPayPalStatusChanged Auth webhooks
 
 ## 2.10.0
 * Bank Account verifications API